new edits
I chose USA as a country, and New Smyrna as a beach because i found that this is the most dangerous beach when it comes to shark attacks in the united states. i sliced data based on beach, year, injury counts, the total injury counts from 1962 til 2024 and the recent injury counts in the last 5 years.